supportconfig: Change method to list swapped processes
I have removed the scriptlet which collects VmSwap of /proc/pid/status.
It turned out to be unreliable. The values are not correct and
sometimes processes are missing or still present even if other methods
(smem, /proc/pid/smaps) report otherwise.

Now I go through /proc/pid/smaps and collet SwapPss. This is more
accurate, but runs a bit slower.