Apply scope limits only when they are non-zero
MemoryLow=0 is the kernel default, therefore we don't need to set the
property in such a case.

TasksMax=0 is invalid input for systemd, so we don't set it.

Purpose of this change is to allow to configure which properties are set
on the scope and adjust for various systemd versions that do not support
all properties.